chore: move trial logs in a trial detail page tab [DET-5410]
